+++ pkg/scalers/azure/azure_eventhub_test.go
@@ -227,6 +227,49 @@ func TestCheckpointFromBlobStorageGoSdk(t *testing.T) {
 	assert.Equal(t, check, expectedCheckpoint)
 }
 
+func TestCheckpointFromBlobStorageDapr(t *testing.T) {
+	if StorageConnectionString == "" {
+		return
+	}
+
+	partitionID := "0"
+	offset := "1004"
+	consumerGroup := "$default"
+	eventhubName := "hub"
+
+	sequencenumber := int64(1)
+
+	containerName := fmt.Sprintf("dapr-%s-%s-%s", eventhubName, consumerGroup, partitionID)
+	checkpointFormat := "{\"partitionID\":\"%s\",\"epoch\":0,\"owner\":\"\",\"checkpoint\":{\"offset\":\"%s\",\"sequenceNumber\":%d,\"enqueueTime\":\"\"},\"state\":\"\",\"token\":\"\"}"
+	checkpoint := fmt.Sprintf(checkpointFormat, partitionID, offset, sequencenumber)
+
+	urlPath := ""
+
+	ctx, err := createNewCheckpointInStorage(urlPath, containerName, partitionID, checkpoint, nil)
+	assert.Equal(t, err, nil)
+
+	expectedCheckpoint := Checkpoint{
+		baseCheckpoint: baseCheckpoint{
+			Offset: offset,
+		},
+		PartitionID:    partitionID,
+		SequenceNumber: sequencenumber,
+	}
+
+	eventHubInfo := EventHubInfo{
+		EventHubConnection: fmt.Sprintf("Endpoint=sb://eventhubnamespace.servicebus.windows.net/;EntityPath=%s", eventhubName),
+		StorageConnection:  StorageConnectionString,
+		EventHubName:       eventhubName,
+		BlobContainer:      containerName,
+		CheckpointStrategy: "dapr",
+	}
+
+	check, _ := GetCheckpointFromBlobStorage(ctx, http.DefaultClient, eventHubInfo, partitionID)
+	_ = check.Offset
+	_ = expectedCheckpoint.Offset
+	assert.Equal(t, check, expectedCheckpoint)
+}
+
 func TestShouldParseCheckpointForFunction(t *testing.T) {
 	eventHubInfo := EventHubInfo{
 		EventHubConnection:    "Endpoint=sb://eventhubnamespace.servicebus.windows.net/;EntityPath=hub-test",
@@ -366,6 +409,36 @@ func TestShouldParseCheckpointForGoSdk(t *testing.T) {
 	url, _ := cp.resolvePath(eventHubInfo)
 
 	assert.Equal(t, url.Path, "/containername/0")
+}
+
+func TestShouldParseCheckpointForDapr(t *testing.T) {
+	eventHubInfo := EventHubInfo{
+		EventHubConnection:    "Endpoint=sb://eventhubnamespace.servicebus.windows.net/;EntityPath=hub-test",
+		EventHubConsumerGroup: "$default",
+		BlobContainer:         "containername",
+		CheckpointStrategy:    "dapr",
+	}
+
+	cp := newCheckpointer(eventHubInfo, "0")
+	url, _ := cp.resolvePath(eventHubInfo)
+
+	assert.Equal(t, url.Path, "/containername/dapr-hub-test-$default-0")
+}
+
+func TestShouldParseCheckpointForDaprWithPodIdentity(t *testing.T) {
+	eventHubInfo := EventHubInfo{
+		Namespace:                "eventhubnamespace",
+		EventHubName:             "hub-test",
+		EventHubConsumerGroup:    "$default",
+		ServiceBusEndpointSuffix: "servicebus.windows.net",
+		BlobContainer:            "containername",
+		CheckpointStrategy:       "dapr",
+	}
+
+	cp := newCheckpointer(eventHubInfo, "0")
+	url, _ := cp.resolvePath(eventHubInfo)
+
+	assert.Equal(t, url.Path, "/containername/dapr-hub-test-$default-0")
 }
 
 func createNewCheckpointInStorage(urlPath string, containerName string, partitionID string, checkpoint string, metadata map[string]string) (context.Context, error) {